Modified who gets an email, when a user comments.
Basically I created a Comment Manager that handles the responsability,
of when to send the email, and to which users are sent.

We added the notion of comment observers. This people are the one
interested in the comment.

* When an external user comments, the comments
observers are the admin and the invited author(if exits).

* When either the admin or the invited author comments, the comment
observer is the one that didnt wrote the comment.